+++ CHANGELOG.md
@@ -16,7 +16,7 @@
 
 ### Fixed
 
-- startup issue on Windows platform
+- startup & routing issues on Windows platform
 - `components` dir name config should be applied correctly in various places
 
 ## [0.15.0] - 2025-03-01